refactor: Rename request DTO packages for better organization
